requirements.txt
google-genai
requests
python-dotenv
google-auth